requirements.txt
asgiref             3.5.0
Django              4.0.3
django-cors-headers 3.11.0
djangorestframework 3.13.1
gunicorn            20.1.0
pip                 22.0.4
pytz                2022.1
setuptools          60.9.3
sqlparse            0.4.2
tzdata              2022.1
wheel               0.37.1
whitenoise          6.0.0
